requirements.txt
dj-database-url==0.4.1
Django==1.10.2
djangorestframework==3.4.7
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
psycopg2==2.6.2
PyJWT==1.4.2
whitenoise==3.2.2
